fix: FeedBookmarkService 토글 방식으로 변경

+++ feed-service/src/main/java/com/feed_service/application/service/FeedBookmarkService.java
@@ -2,13 +2,18 @@ package com.feed_service.application.service;
 
 import com.common.exception.CustomException;
 import com.common.response.ErrorCode;
+import com.common.response.SuccessCode;
 import com.feed_service.domain.model.Feed;
 import com.feed_service.domain.model.FeedBookmark;
 import com.feed_service.domain.repository.FeedBookmarkRepository;
 import com.feed_service.domain.repository.FeedRepository;
+import com.feed_service.presentation.response.FeedBookmarkResponseDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+
 @Service
 @RequiredArgsConstructor
 public class FeedBookmarkService {
@@ -16,21 +21,19 @@ public class FeedBookmarkService {
     private final FeedRepository feedRepository;
     private final FeedBookmarkRepository feedBookmarkRepository;
 
-    @Transactional
-    public void bookmarkFeed(Long feedId, Long userId){
+    public FeedBookmarkResponseDto toggleBookmakr(Long feedId, Long userId) {
 
         Feed feed = feedRepository.findById(feedId)
                 .orElseThrow(() -> new CustomException(ErrorCode.NOT_FOUND));
 
-        if(feedBookmarkRepository.existsByFeed_IdAndUserId(feedId, userId)){
-            throw new CustomException(ErrorCode.CONFLICT);
+        boolean exists = feedBookmarkRepository.existsByFeed_IdAndUserId(feedId, userId);
+
+        if (exists) {
+            feedBookmarkRepository.deleteByFeed_IdAndUserId(feedId, userId);
+        } else{
+            feedBookmarkRepository.save(new FeedBookmark(feed, userId));
         }
 
-        feedBookmarkRepository.save(new  FeedBookmark(feed, userId));
-    }
-
-    @Transactional
-    public void deleteBookmarkFeed(Long feedId, Long userId){
-        feedBookmarkRepository.deleteByFeed_IdAndUserId(feedId, userId);
+        return new FeedBookmarkResponseDto(!exists);
     }
 }
